+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Protocol: `channel_update` sent to disable channel only if we reject an HTLC.
+- Protocol: we don't send redundant `node_announcement` on every new channel.
 
 ### Deprecated
 
@@ -37,6 +38,7 @@ changes.
 - JSON RPC: `getroute` `fuzzpercent` and `pay` `maxfeepercent` can now be > 100.
 - Protocol: fix occasional deadlock when both peers flood with gossip.
 - Protocol: fix occasional long delay on sending `reply_short_channel_ids_end`.
+- Protocol: re-send `node_announcement` when address/alias/color etc change.
 
 ### Security